fix(cgroups): check that the con.pid is not 0
0 means that sway wasn't able to get the pid of an X11 window. It's
highly likely that we'll get the same result with _NET_WM_PID lookup,
but at least the error would point to the right direction.

Note: more complex logic would be required to detect sandboxed windows
(flatpak). Those are already isolated, but the pid we get is valid only
within the sandbox and may belong to a different process in the host
system.