fix: partially revert existing session detection (898b9b75)
The only reliable way to detect another session is to check for
(WAYLAND_)?DISPLAY in the systemd environment, but that results in a
false positive if another command in the config has already set the
variable.

As Sway launches all `exec`s in parallel, it is highly possible that a
shorter scriplet (`exec systemctl --user import-environment`) will be
finished first.