feat: add systemd-localed integration.
A typical Fedora installation will have its locale and keyboard settings
available via `org.freedesktop.locale1` interface and already applied
for Xorg.

It may be convenient to apply the same settings to Sway, allowing to
keep the configuration in a single place and to use tools that interact
with localed (graphical configurators, etc).

There are also scenarios where it's necessary to use the system keyboard
configuration. For example, when Sway is set as a display server for a
greeter and user accounts created from installer require non-default
keyboard layout to login.