feat(session): make cleanup code default
It used to be opt-in as I had doubts about safety and convenience of the
behavior. The default config had `--with-cleanup` and I haven't seen
anyone disabling the option or reporting any problems with it.

BREAKING: now you'll have to opt-out of the session teardown code with
the new option `session.sh --no-cleanup`.