feat(session): more reliable session teardown
`graphical-session.target` can outlive the `sway-session.target` if
there are running services with a strong direct dependency on it.
E.g., anything that `Requires=graphical-session.target` could prevent it
from stopping.

A fancy trick implemented in gnome-session is to start something that
conflicts with both sway- and graphical-session targets and forces these
units to stop.