feat: export XCURSOR_THEME/SIZE
The change isn't really required; most of the applications and frameworks
nowadays have support for server-rendered cursors via cursor-shape-v1,
and the main outlier, GTK, doesn't support the variables either.

And yet it is a part of the environment set by the compositor, so it would be
unfair to not expose it to the graphical apps running as systemd user services.

Note: `swaymsg seat <seat> xcursor_theme ...` would change the values passed to
a new applications, but we won't be able to react to this update.